Add basic handling for nested lists
We've seen a slow but continual stream of errors related to markdown rendering of
nested lists. We'd like to offer at least a basic solution to this common use
case. The output of this change isn't valid markdown, but represents a human-
readable result, which is at least a step forward.